refactor(worker): extract test module into tests.rs (sc-1636)
Move the worker's inline `#[cfg(test)] mod tests` out of lib.rs into a
dedicated src/tests.rs declared via `#[cfg(test)] mod tests;`. `super`
still resolves to the crate root, so the move is behavior-preserving
with no production change. lib.rs drops 6,003 -> 4,854 lines.

Slice of the lib.rs module split (sc-1636); independent of the API
slice (different crate/file).
